apps/api/requirements.txt
# lore-anchor Backend API dependencies
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic-settings>=2.2.0,<3.0.0
boto3>=1.34.0,<2.0.0
redis>=5.0.0,<6.0.0
supabase>=2.4.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
slowapi>=0.1.9,<1.0.0
stripe>=9.0.0,<10.0.0

# --- Dev / CI dependencies ---
ruff>=0.4.0,<1.0.0
mypy>=1.10.0,<2.0.0

# --- Test dependencies ---
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
